requirements.txt
aws_cdk.core
aws_cdk.aws_lambda
aws_cdk.aws_apigateway